int64_t to long long

+++ two_sets_2_/two_sets_2.cpp
@@ -7,9 +7,9 @@
 
 using namespace std;
 
-int64_t check_if_eligible(int64_t n) {
+long long check_if_eligible(long long n) {
     // Lukujen 1...n summa voidaan laskea kaavalla (n*(n+1))/2. Jos summa on pariton, vastaus on NO.
-    int64_t sum = (n * (n + 1)) / 2;
+    long long sum = (n * (n + 1)) / 2;
     if (sum % 2 != 0) {
         cout << 0;
         return -1;
@@ -19,29 +19,29 @@ int64_t check_if_eligible(int64_t n) {
 }
 
 int main() {
-    int64_t n;
+    long long n;
     cin >> n;
-    int64_t modulo = pow(10, 9) + 7;
-    int64_t sum = check_if_eligible(n);
+    long long modulo = pow(10, 9) + 7;
+    long long sum = check_if_eligible(n);
     if (sum == -1) return 0;
-    int64_t answer = sum / 2;
+    long long answer = sum / 2;
     
-    vector<int64_t> numbers;
-    for (int64_t i = 1; i <= n; i++) {
+    vector<long long> numbers;
+    for (long long i = 1; i <= n; i++) {
         numbers.push_back(i);
     }
 
-    vector<vector<int64_t>> sums(n,vector<int64_t>(answer+1,0));
+    vector<vector<long long>> sums(n,vector<long long>(answer+1,0));
     sums[0][0] = 1;
 
-    for (int64_t a = 1; a < n; a++) {
-        for (int64_t b = 0; b <= answer; b++) {
+    for (long long a = 1; a < n; a++) {
+        for (long long b = 0; b <= answer; b++) {
             // Two possibilities: either we include a in the sum or not.
             // 1. We don't include it, and thus 
-            int64_t notIncluded = sums[a-1][b];
+            long long notIncluded = sums[a-1][b];
             // 1. We included it, and thus
-            int64_t included = sums[a-1][b-a];
-            int64_t total = notIncluded;
+            long long included = sums[a-1][b-a];
+            long long total = notIncluded;
             if (b - a >= 0) {
                 total = included + notIncluded;
                 total %= modulo;
